added random for lootBoxFActory

+++ contracts/CyberWayNFT.sol
@@ -7,9 +7,9 @@ import "./Governance.sol";
 contract CyberWayNFT is ERC721, Governance {
 
     struct CyberWayToken {
-        uint256 kind; // 0 - character, 1 - car
-        uint256 colorFrame; // Grey, Green, Blue, Purple, Gold
-        uint256 rand; // 5 Common,Uncommon,Rare,Epic,Legendary
+        uint8 kind; // 0 - character, 1 - car
+        uint8 colorFrame; // Grey, Green, Blue, Purple, Gold
+        uint8 rand; // 5 Common,Uncommon,Rare,Epic,Legendary
     }
 
     CyberWayToken[] public nftTokens;
@@ -18,9 +18,9 @@ contract CyberWayNFT is ERC721, Governance {
 
 
     function mint(address to,
-                    uint256 kind_,
-                    uint256 newColorFrame_,
-                    uint256 rand_) external onlyGovernance returns(uint256){
+                    uint8 kind_,
+                    uint8 newColorFrame_,
+                    uint8 rand_) external onlyGovernance returns(uint256){
         CyberWayToken memory cyberToken = CyberWayToken({kind: kind_, colorFrame: newColorFrame_, rand: rand_});
         nftTokens.push(cyberToken);
         uint256 tokenId =  nftTokens.length - 1;
@@ -35,17 +35,17 @@ contract CyberWayNFT is ERC721, Governance {
     }
 
 
-    function getTokenKind(uint256 tokenId) public view returns(uint256) {
+    function getTokenKind(uint256 tokenId) public view returns(uint8) {
         return nftTokens[tokenId].kind;
     }
 
 
-    function getTokenColor(uint256 tokenId) public view returns(uint256) {
+    function getTokenColor(uint256 tokenId) public view returns(uint8) {
         return nftTokens[tokenId].colorFrame;
     }
 
 
-    function getTokenRand(uint256 tokenId) public view returns(uint256) {
+    function getTokenRand(uint256 tokenId) public view returns(uint8) {
         return nftTokens[tokenId].rand;
     }
 }
